build and install libbam.a, libbcf.a, install header files
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/Rsamtools@53907 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/samtools/ksort.h
@@ -256,7 +256,7 @@ typedef struct {
 		int i, j;														\
 		for (i = n; i > 1; --i) {										\
 			type_t tmp;													\
-			j = (int)(((float) rand()) / RAND_MAX * i);	\
+			j = (int)(((double)rand() / RAND_MAX) * i);	\
 			tmp = a[j]; a[j] = a[i-1]; a[i-1] = tmp;					\
 		}																\
 	}
